Started working on shared codebase for pages
