+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Piral Changelog
 
+## 0.8.1 (tbd)
+
+* Improved default template
+
 ## 0.8.0 (October 21, 2019)
 
 * Use declaration merging instead of generics (#72)
