+++ src/pages/common/components/ScrollToTop.tsx
@@ -1,0 +1,6 @@
+import * as React from 'react';
+
+export const ScrollToTop: React.FC = () => {
+  window.scrollTo(0, 0);
+  return null;
+};
