+++ src/pages/docs/src/index.html
@@ -1,8 +1,8 @@
 <!DOCTYPE html>
 <html lang="en">
 <meta charset="utf-8">
-<title>Documentation - Piral</title>
-<meta http-equiv="X-UA-Compatible" content="IE=edge">
+<title>Piral - Documentation</title>
+<meta http-equiv="X-UA-Compatible" content="ie=edge">
 <meta name="viewport" content="width=device-width, initial-scale=1.0">
 <meta name="description" content="The documentation and guidelines for using Piral - the React-based framework for building microfrontends.">
 <meta name="author" content="smapiot">
@@ -11,7 +11,7 @@
 <div id="app">
   <noscript class="app-center">
     <section class="row">
-      <img src="assets/noscript-robot.png" alt="Noscript Sad Robot">
+      <img src="../../common/assets/noscript-robot.png" alt="Noscript Sad Robot">
       <h2>Sorry!</h2>
       <h3>We require JavaScript to be enabled.</h3>
     </div>
