--- src/pages/docs/src/scripts/components/ScrollToTop.tsx
@@ -1,4 +1,0 @@
-export const ScrollToTop = (): false => {
-  window.scrollTo(0, 0);
-  return false;
-};
