+++ src/pages/docs/src/scripts/components/index.ts
@@ -1,10 +1,7 @@
 export * from './ContentPage';
-export * from './FooterMenu';
 export * from './ImageCard';
-export * from './Loader';
 export * from './Markdown';
 export * from './Menu';
-export * from './ScrollToTop';
 export * from './Search';
 export * from './Sidebar';
 export * from './Tabs';
