+++ src/pages/www/src/index.html
@@ -1,13 +1,21 @@
 <!DOCTYPE html>
 <html lang="en">
-<meta charset="UTF-8">
-<title>Piral.io</title>
-<meta name="viewport" content="width=device-width, initial-scale=1.0, shrink-to-fit=no">
+  <meta charset="utf-8">
+<title>Piral - Portal Solutions using Microfrontends</title>
 <meta http-equiv="X-UA-Compatible" content="ie=edge">
+<meta name="viewport" content="width=device-width, initial-scale=1.0, shrink-to-fit=no">
 <meta name="description" content="Piral - Build next generation portal applications."><meta name="author" content="smapiot">
-<link href="https://fonts.googleapis.com/css?family=Lato:300,400,700,900" rel="stylesheet" type="text/css">
-<link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.0.8/css/all.css">
+<meta name="author" content="smapiot">
+<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.8.2/css/all.min.css">
 <link rel="stylesheet" href="./style/site.scss">
-<div id="app"></div>
+<div id="app">
+  <noscript class="app-center">
+    <section class="row">
+      <img src="../../common/assets/noscript-robot.png" alt="Noscript Sad Robot">
+      <h2>Sorry!</h2>
+      <h3>We require JavaScript to be enabled.</h3>
+    </div>
+  </noscript>
+</div>
 <script src="scripts/index.tsx"></script>
 </html>
