+++ src/pages/www/src/style/_custom.scss
@@ -147,7 +147,7 @@ button {
   }
 
   .fa-heart {
-    color: $heart-color;
+    color: $color-heart;
   }
 }
