+++ src/pages/www/src/style/site.scss
@@ -1,7 +1,4 @@
 @import "../../../../../node_modules/bootstrap/scss/bootstrap";
-
-$color-primary: #00FFAD;
-$color-secondary: #192440;
-$heart-color: #ea5395;
-
+@import "../../../common/styles/variables";
 @import "custom";
+@import "../../../common/styles/footer";
